requirements.txt
gtts==2.3.1
pydub==0.25.1
openai==1.1.0
python-dotenv==1.0.0